+++ src/main/java/net/hackyourfuture/coursehub/data/InstructorEntity.java
@@ -1,7 +1,7 @@
 package net.hackyourfuture.coursehub.data;
 
 public record InstructorEntity(
-        Long instructorId,
+        Integer instructorId,
         String firstName,
         String lastName,
         String email
