+++ src/main/java/net/hackyourfuture/coursehub/repository/EnrollmentRepository.java
@@ -12,9 +12,9 @@ import java.util.Map;
 @Repository
 public class EnrollmentRepository {
     public static final RowMapper<EnrollmentEntity> ENROLLMENT_ROW_MAPPER = (rs, rowNum) -> new EnrollmentEntity(
-            rs.getLong("enrollment_id"),
-            rs.getLong("course_id"),
-            rs.getLong("student_id"),
+            rs.getInt("enrollment_id"),
+            rs.getInt("course_id"),
+            rs.getInt("student_id"),
             rs.getObject("enrollment_date", LocalDate.class)
     );
     private final NamedParameterJdbcTemplate jdbcTemplate;
@@ -23,12 +23,12 @@ public class EnrollmentRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public List<EnrollmentEntity> findEnrollmentsByCourseId(Long courseId) {
+    public List<EnrollmentEntity> findEnrollmentsByCourseId(Integer courseId) {
         String sql = "SELECT * FROM enrollment WHERE course_id = :courseId";
         return jdbcTemplate.query(sql, Map.of("courseId", courseId), ENROLLMENT_ROW_MAPPER);
     }
 
-    public List<EnrollmentEntity> findEnrollmentsByStudentId(Long studentId) {
+    public List<EnrollmentEntity> findEnrollmentsByStudentId(Integer studentId) {
         String sql = "SELECT * FROM enrollment WHERE student_id = :studentId";
         return jdbcTemplate.query(sql, Map.of("studentId", studentId), ENROLLMENT_ROW_MAPPER);
     }
