+++ src/main/java/net/hackyourfuture/coursehub/repository/StudentRepository.java
@@ -10,7 +10,7 @@ import java.util.Map;
 @Repository
 public class StudentRepository {
     public static final RowMapper<StudentEntity> STUDENT_ROW_MAPPER = (rs, rowNum) -> new StudentEntity(
-            rs.getLong("student_id"),
+            rs.getInt("student_id"),
             rs.getString("first_name"),
             rs.getString("last_name"),
             rs.getString("email")
@@ -21,7 +21,7 @@ public class StudentRepository {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public StudentEntity findById(Long studentId) {
+    public StudentEntity findById(Integer studentId) {
         String sql = "SELECT * FROM student WHERE student_id = :studentId";
         return jdbcTemplate.queryForObject(sql, Map.of("studentId", studentId), STUDENT_ROW_MAPPER);
     }
